requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
SQLAlchemy==2.0.43
asyncpg==0.29.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
psycopg[binary]==3.2.9
email-validator==2.2.0
python-dotenv==1.0.1
bcrypt==3.2.2
